refactor: 1) new design 2) remove unnecessary files

+++ src/main/java/entity/User.java
@@ -114,7 +114,7 @@ public class User  {
     public User() {
     }
 
-    public User(int id) {
+    public User(long id) {
         this.id = id;
     }
